243: move Block
