+++ src/rpc_client/mod.rs
@@ -1,5 +1,5 @@
 use crate::rpc_client::eth_rpc::{
-    are_errors_consistent, Block, BlockSpec, FeeHistory, FeeHistoryParams, GetBlockByNumberParams,
+    are_errors_consistent, BlockSpec, FeeHistory, FeeHistoryParams, GetBlockByNumberParams,
     Hash, HttpResponsePayload, ResponseSizeEstimate, SendRawTransactionResult,
     HEADER_SIZE_LIMIT,
 };
@@ -14,7 +14,7 @@ use requests::GetLogsParam;
 use serde::{de::DeserializeOwned, Serialize};
 use std::collections::BTreeMap;
 use std::fmt::Debug;
-use responses::LogEntry;
+use responses::{Block, LogEntry};
 
 pub mod checked_amount;
 pub(crate) mod eth_rpc;
