+++ src/rpc_client/responses.rs
@@ -2,7 +2,7 @@ use crate::rpc_client::checked_amount::CheckedAmountOf;
 use crate::rpc_client::eth_rpc::{
     Data, FixedSizeData, Hash, HttpResponsePayload, ResponseTransform,
 };
-use crate::rpc_client::numeric::{BlockNumber, GasAmount, LogIndex, WeiPerGas};
+use crate::rpc_client::numeric::{BlockNumber, GasAmount, LogIndex, Wei, WeiPerGas};
 use ic_ethereum_types::Address;
 use serde::{Deserialize, Serialize};
 use std::fmt::{Display, Formatter};
@@ -152,5 +152,51 @@ pub struct LogEntry {
 impl HttpResponsePayload for Vec<LogEntry> {
     fn response_transform() -> Option<ResponseTransform> {
         Some(ResponseTransform::LogEntries)
+    }
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+pub struct Block {
+    #[serde(rename = "baseFeePerGas")]
+    pub base_fee_per_gas: Option<Wei>,
+    pub number: BlockNumber,
+    pub difficulty: Option<CheckedAmountOf<()>>,
+    #[serde(rename = "extraData")]
+    pub extra_data: String,
+    #[serde(rename = "gasLimit")]
+    pub gas_limit: CheckedAmountOf<()>,
+    #[serde(rename = "gasUsed")]
+    pub gas_used: CheckedAmountOf<()>,
+    pub hash: String,
+    #[serde(rename = "logsBloom")]
+    pub logs_bloom: String,
+    pub miner: String,
+    #[serde(rename = "mixHash")]
+    pub mix_hash: String,
+    pub nonce: CheckedAmountOf<()>,
+    #[serde(rename = "parentHash")]
+    pub parent_hash: String,
+    #[serde(rename = "receiptsRoot")]
+    pub receipts_root: String,
+    #[serde(rename = "sha3Uncles")]
+    pub sha3_uncles: String,
+    pub size: CheckedAmountOf<()>,
+    #[serde(rename = "stateRoot")]
+    pub state_root: String,
+    #[serde(rename = "timestamp")]
+    pub timestamp: CheckedAmountOf<()>,
+    #[serde(rename = "totalDifficulty")]
+    pub total_difficulty: Option<CheckedAmountOf<()>>,
+    #[serde(default)]
+    pub transactions: Vec<String>,
+    #[serde(rename = "transactionsRoot")]
+    pub transactions_root: Option<String>,
+    #[serde(default)]
+    pub uncles: Vec<String>,
+}
+
+impl HttpResponsePayload for Block {
+    fn response_transform() -> Option<ResponseTransform> {
+        Some(ResponseTransform::Block)
     }
 }
